Secado Sistema Automatizado/requirements.txt
# Azure Functions
azure-functions>=1.13.0

# Autenticación Google
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0

# Procesamiento de datos
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.10.0

# Visualización
matplotlib>=3.8.0
seaborn>=0.13.0

# Machine Learning
scikit-learn>=1.3.0
catboost>=1.2.2

# HTTP requests
requests>=2.31.0

# Configuración
python-dotenv>=1.0.0
pyyaml>=6.0

